+++ patientMatcher/server/controllers.py
@@ -33,7 +33,7 @@ def heartbeat(disclaimer):
 
     hbeat = {
         "heartbeat": {
-            "production": True,
+            "production": bool(current_app.config.get("TESTING") is False),
             "version": __version__,
             "accept": [
                 "application/vnd.ga4gh.matchmaker.v1.0+json",