+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Restore saving of all matching attempts into database, also those with no results
 - Parsing of admins emails list when provided using env vars
+- Heartbeat endpoint to return True or False if the app is prod or staging app
 
 ## [2.13] - 2021-12-13
 ### added